This will finish the recording process for a user
This will actually overwrite the users lircd.conf file. The file that is
going to be copied is in /lirc_node/irrecord/lircd.conf

followed up by a restart of lirc to get the new command into lirc.
Lirc_web will be updated with a function that reloads lirc_node for
these lircd.conf changes to take effect.